[XPU] Add docs for gm_default_size and fix type (#1962)
Add description of gm_default_size and change kunlunxin_gm_default_size‘s type to int64_t

+++ c_api/fastdeploy_capi/runtime/runtime_option.cc
@@ -108,7 +108,7 @@ void FD_C_RuntimeOptionWrapperUseKunlunXin(
                                std::string(precision), 
                                bool(adaptive_seqlen),
                                bool(enable_multi_stream),
-                               gm_default_size);
+                               int64_t(gm_default_size));
 }
 
 void FD_C_RuntimeOptionWrapperUseSophgo(
